add createFeeAssetItem per direction
