+++ src/createXcmCalls/limitedReserveTransferAssets.ts
@@ -2,12 +2,11 @@
 
 import type { ApiPromise } from '@polkadot/api';
 import type { SubmittableExtrinsic } from '@polkadot/api/submittable/types';
+import { u32 } from '@polkadot/types';
 import type { ISubmittableResult } from '@polkadot/types/types';
 
 import { createXcmTypes } from '../createXcmTypes';
-import { createSystemToParaMultiAssets } from '../createXcmTypes/util/createSystemToParaMultiAssets';
 import { Direction } from '../types';
-import { getFeeAssetItemIndex } from '../util/getFeeAssetItemIndex';
 import { normalizeArrToStr } from '../util/normalizeArrToStr';
 import { establishXcmPallet } from './util/establishXcmPallet';
 
@@ -49,21 +48,16 @@ export const limitedReserveTransferAssets = (
 	);
 	const weightLimitType = typeCreator.createWeightLimit(api, weightLimit);
 
-	let feeAssetItem = 0;
-	if (
-		paysWithFeeDest &&
-		xcmVersion === 3 &&
-		direction === Direction.SystemToPara
-	) {
-		const multiAssets = createSystemToParaMultiAssets(
-			api,
-			normalizeArrToStr(amounts),
-			specName,
-			assetIds
-		);
-
-		feeAssetItem = getFeeAssetItemIndex(paysWithFeeDest, multiAssets, specName);
-	}
+	const feeAssetItem: u32 = paysWithFeeDest
+		? typeCreator.createFeeAssetItem(
+				api,
+				paysWithFeeDest,
+				specName,
+				assetIds,
+				amounts,
+				xcmVersion
+		  )
+		: api.registry.createType('u32', 0);
 
 	return ext(dest, beneficiary, assets, feeAssetItem, weightLimitType);
 };
