+++ src/createXcmCalls/reserveTransferAssets.ts
@@ -5,9 +5,7 @@ import type { SubmittableExtrinsic } from '@polkadot/api/submittable/types';
 import type { ISubmittableResult } from '@polkadot/types/types';
 
 import { createXcmTypes } from '../createXcmTypes';
-import { createSystemToParaMultiAssets } from '../createXcmTypes/util/createSystemToParaMultiAssets';
 import { Direction } from '../types';
-import { getFeeAssetItemIndex } from '../util/getFeeAssetItemIndex';
 import { normalizeArrToStr } from '../util/normalizeArrToStr';
 import { establishXcmPallet } from './util/establishXcmPallet';
 
@@ -46,21 +44,16 @@ export const reserveTransferAssets = (
 		assetIds
 	);
 
-	let feeAssetItem = 0;
-	if (
-		paysWithFeeDest &&
-		xcmVersion === 3 &&
-		direction === Direction.SystemToPara
-	) {
-		const multiAssets = createSystemToParaMultiAssets(
-			api,
-			normalizeArrToStr(amounts),
-			specName,
-			assetIds
-		);
-
-		feeAssetItem = getFeeAssetItemIndex(paysWithFeeDest, multiAssets, specName);
-	}
+	const feeAssetItem = paysWithFeeDest
+		? typeCreator.createFeeAssetItem(
+				api,
+				paysWithFeeDest,
+				specName,
+				assetIds,
+				amounts,
+				xcmVersion
+		  )
+		: api.registry.createType('u32', 0);
 
 	return ext(dest, beneficiary, assets, feeAssetItem);
 };
