+++ src/createXcmCalls/teleportAssets.ts
@@ -28,7 +28,8 @@ export const teleportAssets = (
 	amounts: string[],
 	destChainId: string,
 	xcmVersion: number,
-	specName: string
+	specName: string,
+	paysWithFeeDest?: string
 ): SubmittableExtrinsic<'promise', ISubmittableResult> => {
 	const pallet = establishXcmPallet(api);
 	const ext = api.tx[pallet].teleportAssets;
@@ -43,7 +44,9 @@ export const teleportAssets = (
 		assetIds
 	);
 
-	const feeAssetItem = 0;
+	const feeAssetItem = paysWithFeeDest
+		? typeCreator.createFeeAssetItem(api)
+		: api.registry.createType('u32', 0);
 
 	return ext(dest, beneficiary, assets, feeAssetItem);
 };
