+++ src/createXcmTypes/RelayToPara.ts
@@ -1,6 +1,7 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
 import type { ApiPromise } from '@polkadot/api';
+import { u32 } from '@polkadot/types';
 import type {
 	MultiAssetsV2,
 	VersionedMultiAssets,
@@ -155,5 +156,11 @@ export const RelayToPara: ICreateXcmType = {
 			: { Unlimited: null };
 
 		return api.registry.createType('XcmV2WeightLimit', limit);
+	},
+
+	createFeeAssetItem: (api: ApiPromise): u32 => {
+		const feeAssetItem = api.registry.createType('u32', 0);
+
+		return feeAssetItem;
 	},
 };
