+++ src/createXcmTypes/SystemToPara.ts
@@ -1,6 +1,7 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
 import type { ApiPromise } from '@polkadot/api';
+import { u32 } from '@polkadot/types';
 import type {
 	MultiAssetsV2,
 	VersionedMultiAssets,
@@ -10,8 +11,16 @@ import type {
 import type { XcmV3MultiassetMultiAssets } from '@polkadot/types/lookup';
 import { isEthereumAddress } from '@polkadot/util-crypto';
 
+import { findRelayChain, parseRegistry } from '../registry';
+import { getFeeAssetItemIndex } from '../util/getFeeAssetItemIndex';
+import { normalizeArrToStr } from '../util/normalizeArrToStr';
+import { MultiAsset, MultiAssetInterior } from './../types';
 import { ICreateXcmType, IWeightLimit } from './types';
-import { createSystemToParaMultiAssets } from './util/createSystemToParaMultiAssets';
+import { isAscendingOrder } from './util/checkIsAscendingOrder';
+import { dedupeMultiAssets } from './util/dedupeMultiAssets';
+import { fetchPalletInstanceId } from './util/fetchPalletInstanceId';
+import { getSystemChainTokenSymbolGeneralIndex } from './util/getTokenSymbolGeneralIndex';
+import { sortMultiAssetsAscending } from './util/sortMultiAssetsAscending';
 
 export const SystemToPara: ICreateXcmType = {
 	/**
@@ -150,4 +159,113 @@ export const SystemToPara: ICreateXcmType = {
 
 		return api.registry.createType('XcmV2WeightLimit', limit);
 	},
+
+	createFeeAssetItem: (
+		api: ApiPromise,
+		paysWithFeeDest?: string,
+		specName?: string,
+		assetIds?: string[],
+		amounts?: string[],
+		xcmVersion?: number
+	): u32 => {
+		if (
+			xcmVersion &&
+			xcmVersion === 3 &&
+			specName &&
+			amounts &&
+			assetIds &&
+			paysWithFeeDest
+		) {
+			const multiAssets = createSystemToParaMultiAssets(
+				api,
+				normalizeArrToStr(amounts),
+				specName,
+				assetIds
+			);
+
+			const assetIndex = getFeeAssetItemIndex(
+				paysWithFeeDest,
+				multiAssets,
+				specName
+			);
+			const feeAssetItem = api.registry.createType('u32', assetIndex);
+
+			return feeAssetItem;
+		}
+
+		return api.registry.createType('u32', 0);
+	},
+};
+
+/**
+ * Creates and returns a MultiAsset array for system parachains based on provided specName, assets and amounts
+ *
+ * @param api ApiPromise[]
+ * @param amounts string[]
+ * @param specName string
+ * @param assets string[]
+ */
+export const createSystemToParaMultiAssets = (
+	api: ApiPromise,
+	amounts: string[],
+	specName: string,
+	assets: string[]
+): MultiAsset[] => {
+	const palletId = fetchPalletInstanceId(api);
+	const multiAssets = [];
+	const registry = parseRegistry({});
+	const relayChain = findRelayChain(specName, registry);
+
+	// We know this is a System parachain direction which is chainId 1000.
+	const { tokens } = registry[relayChain]['1000'];
+
+	for (let i = 0; i < assets.length; i++) {
+		let assetId: string = assets[i];
+		const amount = amounts[i];
+
+		const parsedAssetIdAsNumber = Number.parseInt(assetId);
+		const isNotANumber = Number.isNaN(parsedAssetIdAsNumber);
+		const isRelayNative = isRelayNativeAsset(tokens, assetId);
+
+		if (!isRelayNative && isNotANumber) {
+			assetId = getSystemChainTokenSymbolGeneralIndex(assetId, specName);
+		}
+
+		const interior: MultiAssetInterior = isRelayNative
+			? { Here: '' }
+			: { X2: [{ PalletInstance: palletId }, { GeneralIndex: assetId }] };
+		const parents = isRelayNative ? 1 : 0;
+
+		const multiAsset = {
+			id: {
+				Concrete: {
+					parents,
+					interior,
+				},
+			},
+			fun: {
+				Fungible: amount,
+			},
+		};
+
+		multiAssets.push(multiAsset);
+	}
+
+	if (!isAscendingOrder(multiAssets)) {
+		sortMultiAssetsAscending(multiAssets);
+	}
+
+	const sortedAndDedupedMultiAssets = dedupeMultiAssets(multiAssets);
+
+	return sortedAndDedupedMultiAssets;
+};
+
+const isRelayNativeAsset = (tokens: string[], assetId: string): boolean => {
+	for (const token of tokens) {
+		if (token.toLowerCase() === assetId.toLowerCase()) {
+			return true;
+		}
+	}
+
+	return false;
 };
